requirements.txt
PyYAML==6.0.2
reportlab==4.2.2
PyMuPDF==1.24.9
requests==2.32.3
